Initial upload
Simple Sign In app that allows you to create a user with a password and then display it on a list.  Some checks are put in too; for repeated user names, repeated sequences in passwords.  It just uses a basic text file to read and write from.  Will probably switch to a sqlite database later if I feel like working on it more.  This provides all the files so you should be able to download and run to install.